Fix compile warnings (closes #23)

+++ plugins/newae/tnewaedevice.h
@@ -68,7 +68,7 @@ protected:
     const QString READ_ONLY_STRING = "alwaysRunFunc";
     const QString WRITE_ONLY_STRING = "writeOnlyFunc";
 
-    const size_t TIMER_READ_INTERVAL = 250;
+    const qint64 TIMER_READ_INTERVAL = 250;
     QByteArray m_readBuffer;       // dynamic buffer
     QElapsedTimer m_lastReadTimer;
     QMutex m_readMutex;
